pyproject.toml
[tool.poetry]
name = "zibal-entrance-project"
version = "0.1.0"
description = ""
authors = ["shervin <shervinsaeedi78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
mongoengine = "^0.23.1"
django-rest-framework-mongoengine = "^3.4.1"
django-rest-framework = "^0.1.0"


[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
